style(widgetDesc): factorize common style with column description
